feat:Uniform return type

+++ zy-micro-service-apps/zy-micro-service-common/src/main/java/com/zy/apps/common/domain/vo/IErrorCode.java
@@ -5,7 +5,7 @@ package com.zy.apps.common.domain.vo;
  * Created by macro on 2019/4/19.
  */
 public interface IErrorCode {
-    long getCode();
+    Integer getCode();
 
     String getMessage();
 }
